Add files via upload
This is the content in the new column for KAE.txt which enables scripts. The scripts enable a new item on the right click menu when a card with a script is selected. The scipt will spawn a card and deuct force/and or build costs from the player. The messgae console will show these transactions indivisually and a comment messgae explaining what has happened and which card triggered it. 

The scripts have been tested in Lackey by me.